Fix pep425 reporting on recent macos versions
On recent macos python might report 10.16 if it was built with an
older sdk. So we try to get the real version by disabling the
SYSTEM_VERSION_COMPAT option and see if we get something different.
Note that we also reset the minor version to 0 if we are on
macos >= 11 to follow the behaviour of the "packaging" library