Add Alpine Linux as platform
[Alpine Linux][1] is a small, simple and secure Linux distro that uses
musl (instead of the more common glibc) as standard C lib, OpenRC
as init system, and the Alpine Package Keeper (apk) as package manager.

Alpine is also introduced as a family, as it fundamentally different
to both Debian and Red Hat families of Linux distros. And in fact,
with Postmarket OS a notable other Linux distro is based directly on
Alpine, so there is in fact a (small) family of distros based on Alpine.

[1]: https://alpinelinux.org/